+++ crates/core/guest/src/memory.rs
@@ -4,10 +4,10 @@ use selium_abi::{
 
 use crate::{GuestError, Result, hostcall::hostcall_ready};
 
-const SHARED_REGION_HEADER_CAPACITY_OFFSET: u32 = 8;
-const SHARED_REGION_HEADER_COUNT_OFFSET: u32 = 16;
-const SHARED_REGION_HEADER_ENTRY_OFFSET: u32 = 24;
-const SHARED_REGION_HEADER_ENTRY_SIZE: u32 = 8;
+const SHARED_REGION_HEADER_CAPACITY_OFFSET: u64 = 8;
+const SHARED_REGION_HEADER_COUNT_OFFSET: u64 = 16;
+const SHARED_REGION_HEADER_ENTRY_OFFSET: u64 = 24;
+const SHARED_REGION_HEADER_ENTRY_SIZE: u64 = 8;
 /// Magic value for multi-memory shared region layout headers.
 pub const SHARED_REGION_MAGIC: u64 = 0x53454C49554D454D;
 
@@ -29,14 +29,14 @@ pub struct SharedMemory {
 /// The region header records `memory_count` and `(offset, len)` pairs.
 /// After `seal()`, no further modifications are permitted.
 pub struct SharedRegionBuilder {
-    capacity: u32,
-    memories: Vec<u32>,
+    capacity: u64,
+    memories: Vec<u64>,
     sealed: bool,
 }
 
 impl SharedRegion {
     /// Allocates a shared memory region with the requested size and alignment.
-    pub fn allocate(size: u32, alignment: u32) -> Result<Self> {
+    pub fn allocate(size: u64, alignment: u64) -> Result<Self> {
         match hostcall_ready(HostcallRequest::SharedMemoryAllocate { size, alignment })? {
             HostcallOutput::SharedRegion(descriptor) => Ok(Self { descriptor }),
             _ => Err(GuestError::UnexpectedHostcallOutput),
@@ -54,7 +54,7 @@ impl SharedRegion {
     }
 
     /// Returns the region length in bytes.
-    pub fn len(&self) -> u32 {
+    pub fn len(&self) -> u64 {
         self.descriptor.len
     }
 
@@ -64,14 +64,14 @@ impl SharedRegion {
     }
 
     /// Attaches to an existing shared region by its shared id and length.
-    pub fn attach(shared_id: u64, len: u32) -> Self {
+    pub fn attach(shared_id: u64, len: u64) -> Self {
         Self {
             descriptor: SharedRegionDescriptor { shared_id, len },
         }
     }
 
     /// Reads the number of sub-memories from the region layout header.
-    pub fn memory_count(&self) -> Result<u32> {
+    pub fn memory_count(&self) -> Result<usize> {
         let mapping = SharedMemory::attach_shared(self.shared_id(), 0, self.len())?;
         let count = mapping.memory_count()?;
         mapping.detach()?;
@@ -79,7 +79,7 @@ impl SharedRegion {
     }
 
     /// Reads the offset and length of the sub-memory at the given index.
-    pub fn memory(&self, index: u32) -> Result<(u32, u32)> {
+    pub fn memory(&self, index: usize) -> Result<(u64, u64)> {
         let mapping = SharedMemory::attach_shared(self.shared_id(), 0, self.len())?;
         let result = mapping.memory(index)?;
         mapping.detach()?;
@@ -99,12 +99,12 @@ impl SharedRegion {
 
 impl SharedMemory {
     /// Attaches a local mapping to a shared region descriptor.
-    pub fn attach(region: SharedRegionDescriptor, offset: u32, len: u32) -> Result<Self> {
+    pub fn attach(region: SharedRegionDescriptor, offset: u64, len: u64) -> Result<Self> {
         Self::attach_shared(region.shared_id, offset, len)
     }
 
     /// Attaches a local mapping to a shared region id.
-    pub fn attach_shared(shared_id: u64, offset: u32, len: u32) -> Result<Self> {
+    pub fn attach_shared(shared_id: u64, offset: u64, len: u64) -> Result<Self> {
         match hostcall_ready(HostcallRequest::SharedMemoryAttach {
             shared_id,
             offset,
@@ -131,7 +131,7 @@ impl SharedMemory {
     }
 
     /// Reads bytes from the mapping at the supplied offset.
-    pub fn read(&self, offset: u32, len: u32) -> Result<Vec<u8>> {
+    pub fn read(&self, offset: u64, len: u64) -> Result<Vec<u8>> {
         match hostcall_ready(HostcallRequest::SharedMemoryRead {
             local_id: self.descriptor.local_id,
             offset,
@@ -143,7 +143,7 @@ impl SharedMemory {
     }
 
     /// Writes bytes into the mapping at the supplied offset.
-    pub fn write(&self, offset: u32, bytes: Vec<u8>) -> Result<()> {
+    pub fn write(&self, offset: u64, bytes: Vec<u8>) -> Result<()> {
         match hostcall_ready(HostcallRequest::SharedMemoryWrite {
             local_id: self.descriptor.local_id,
             offset,
@@ -155,7 +155,7 @@ impl SharedMemory {
     }
 
     /// Atomically adds to a little-endian `u64` at the supplied mapping offset.
-    pub fn fetch_add_u64(&self, offset: u32, value: u64) -> Result<u64> {
+    pub fn fetch_add_u64(&self, offset: u64, value: u64) -> Result<u64> {
         match hostcall_ready(HostcallRequest::SharedMemoryFetchAddU64 {
             local_id: self.descriptor.local_id,
             offset,
@@ -167,7 +167,7 @@ impl SharedMemory {
     }
 
     /// Atomically compares and exchanges a little-endian `u64` at the supplied mapping offset.
-    pub fn compare_exchange_u64(&self, offset: u32, current: u64, new: u64) -> Result<u64> {
+    pub fn compare_exchange_u64(&self, offset: u64, current: u64, new: u64) -> Result<u64> {
         match hostcall_ready(HostcallRequest::SharedMemoryCompareExchangeU64 {
             local_id: self.descriptor.local_id,
             offset,
@@ -190,9 +190,9 @@ impl SharedMemory {
     }
 
     /// Reads the number of sub-memories from the shared region layout header.
-    pub fn memory_count(&self) -> Result<u32> {
+    pub fn memory_count(&self) -> Result<usize> {
         let bytes = self.read(SHARED_REGION_HEADER_COUNT_OFFSET, 4)?;
-        let count = u32::from_le_bytes(
+        let count = usize::from_le_bytes(
             bytes
                 .try_into()
                 .map_err(|_error| GuestError::Host("invalid memory count".to_string()))?,
@@ -201,21 +201,21 @@ impl SharedMemory {
     }
 
     /// Reads the offset and length of the sub-memory at the given index.
-    pub fn memory(&self, index: u32) -> Result<(u32, u32)> {
+    pub fn memory(&self, index: usize) -> Result<(u64, u64)> {
         let count = self.memory_count()?;
         if index >= count {
             return Err(GuestError::Host("memory index out of bounds".to_string()));
         }
         let entry_offset =
-            SHARED_REGION_HEADER_ENTRY_OFFSET + index * SHARED_REGION_HEADER_ENTRY_SIZE;
+            SHARED_REGION_HEADER_ENTRY_OFFSET + index as u64 * SHARED_REGION_HEADER_ENTRY_SIZE;
         let offset_bytes = self.read(entry_offset, 4)?;
         let len_bytes = self.read(entry_offset + 4, 4)?;
-        let offset = u32::from_le_bytes(
+        let offset = u64::from_le_bytes(
             offset_bytes
                 .try_into()
                 .map_err(|_error| GuestError::Host("invalid memory offset".to_string()))?,
         );
-        let len = u32::from_le_bytes(
+        let len = u64::from_le_bytes(
             len_bytes
                 .try_into()
                 .map_err(|_error| GuestError::Host("invalid memory length".to_string()))?,
@@ -226,7 +226,7 @@ impl SharedMemory {
 
 impl SharedRegionBuilder {
     /// Creates a new builder with the given total region capacity.
-    pub fn new(capacity: u32) -> Self {
+    pub fn new(capacity: u64) -> Self {
         Self {
             capacity,
             memories: Vec::new(),
@@ -235,7 +235,7 @@ impl SharedRegionBuilder {
     }
 
     /// Adds a sub-memory of the given length to the layout.
-    pub fn add_memory(&mut self, len: u32) -> Result<&mut Self> {
+    pub fn add_memory(&mut self, len: u64) -> Result<&mut Self> {
         if self.sealed {
             return Err(GuestError::BuilderSealed);
         }
@@ -250,7 +250,7 @@ impl SharedRegionBuilder {
             return Err(GuestError::BuilderSealed);
         }
 
-        let header_size = Self::header_size(self.memories.len() as u32);
+        let header_size = Self::header_size(self.memories.len());
         let mut total = header_size;
         for &len in &self.memories {
             total = Self::align_up(total, 8) + len;
@@ -266,7 +266,7 @@ impl SharedRegionBuilder {
         mapping.write(0, SHARED_REGION_MAGIC.to_le_bytes().to_vec())?;
         mapping.write(
             SHARED_REGION_HEADER_CAPACITY_OFFSET,
-            (self.capacity as u64).to_le_bytes().to_vec(),
+            (self.capacity).to_le_bytes().to_vec(),
         )?;
         mapping.write(
             SHARED_REGION_HEADER_COUNT_OFFSET,
@@ -281,7 +281,7 @@ impl SharedRegionBuilder {
         for (i, &len) in self.memories.iter().enumerate() {
             offset = Self::align_up(offset, 8);
             let entry_offset =
-                SHARED_REGION_HEADER_ENTRY_OFFSET + i as u32 * SHARED_REGION_HEADER_ENTRY_SIZE;
+                SHARED_REGION_HEADER_ENTRY_OFFSET + i as u64 * SHARED_REGION_HEADER_ENTRY_SIZE;
             mapping.write(entry_offset, offset.to_le_bytes().to_vec())?;
             mapping.write(entry_offset + 4, len.to_le_bytes().to_vec())?;
             offset += len;
@@ -291,11 +291,11 @@ impl SharedRegionBuilder {
         Ok(region)
     }
 
-    fn header_size(memory_count: u32) -> u32 {
-        SHARED_REGION_HEADER_ENTRY_OFFSET + memory_count * SHARED_REGION_HEADER_ENTRY_SIZE
+    fn header_size(memory_count: usize) -> u64 {
+        SHARED_REGION_HEADER_ENTRY_OFFSET + memory_count as u64 * SHARED_REGION_HEADER_ENTRY_SIZE
     }
 
-    fn align_up(value: u32, alignment: u32) -> u32 {
+    fn align_up(value: u64, alignment: u64) -> u64 {
         let rem = value % alignment;
         if rem == 0 {
             value
